moved app.models -> app.weather_models
