+++ api/app/tests/weather_models/CMC_glb_RH_TGL_2_latlon.15x.15_2020071300_P000.grib2
