+++ api/app/tests/weather_models/CMC_glb_latlon.15x.15_projection_wkt.txt
