+++ api/app/tests/weather_models/CMC_hrdps_continental_RH_TGL_2_ps2.5km_2020100700_P007-00.grib2
