+++ api/app/tests/weather_models/CMC_hrdps_continental_ps2.5km_projection_wkt.txt
