+++ api/app/tests/weather_models/CMC_reg_RH_TGL_2_ps10km_2020110500_P034.grib2
