+++ api/app/tests/weather_models/test_env_canada_rdps.py
@@ -15,13 +15,13 @@ from pytest_mock import MockerFixture
 import app.time_utils as time_utils
 import app.db.database
 from app.schemas import WeatherStation, Season
-from app.models import env_canada, machine_learning
+from app.weather_models import env_canada, machine_learning
 from app.db.models import (PredictionModel, ProcessedModelRunUrl, PredictionModelRunTimestamp,
                            PredictionModelGridSubset, ModelRunGridSubsetPrediction)
-from app.tests.models.crud import get_actuals_left_outer_join_with_predictions
-from app.tests.models.test_env_canada_gdps import (MockResponse, mock_get_stations,
-                                                   mock_get_model_run_predictions_for_grid,
-                                                   mock_get_actuals_left_outer_join_with_predictions)
+from app.tests.weather_models.crud import get_actuals_left_outer_join_with_predictions
+from app.tests.weather_models.test_env_canada_gdps import (MockResponse, mock_get_stations,
+                                                           mock_get_model_run_predictions_for_grid,
+                                                           mock_get_actuals_left_outer_join_with_predictions)
 # pylint: disable=unused-argument, redefined-outer-name
