+++ api/app/weather_models/machine_learning.py
@@ -9,7 +9,7 @@ from sklearn.linear_model import LinearRegression
 from scipy.interpolate import griddata
 import numpy as np
 from sqlalchemy.orm import Session
-from app.models import MODEL_VALUE_KEYS, construct_interpolated_noon_prediction
+from app.weather_models import MODEL_VALUE_KEYS, construct_interpolated_noon_prediction
 from app.db.models import (
     PredictionModel, PredictionModelGridSubset, ModelRunGridSubsetPrediction)
 from app.db.models.observations import HourlyActual
