+++ openshift/templates/env_canada_hrdps.cronjob.yaml
@@ -46,7 +46,8 @@ objects:
                 - name: ${JOB_NAME}
                   image: docker-registry.default.svc:5000/${PROJECT_TOOLS}/${NAME}-${SUFFIX}:${SUFFIX}
                   imagePullPolicy: "Always"
-                  command: ["python", "-m", "app.models.env_canada", "HRDPS"]
+                  command:
+                    ["python", "-m", "app.weather_models.env_canada", "HRDPS"]
                   env:
                     - name: POSTGRES_READ_USER
                       value: ${NAME}-${SUFFIX}
